refactor(clipboard): split drag-drop readers into partial
Move ReadContentIntoPackage, GetFileDropList and their private
GetText/GetBitmap helpers -- the only Win32ClipboardExtension members
consumed by Win32DragDropExtension -- into
Win32ClipboardExtension.DragDrop.cs. Pure move, verified byte-identical;
also drops the now-unused Windows.Storage and Windows.Win32.UI.Shell
usings from the original file.
